refactor: :recycle: Remove meteor functions from `loadMissedMessages.ts`
Remove Meteor.startup and Tracker.autorun from the `loadMissedMessages.ts` file and implemented a useRef to avoid unnecessary re-rendering.
